+++ apps/docs/components/docs/platform/switcher.tsx
@@ -24,7 +24,7 @@ import {
   usePlatform,
 } from "./context";
 import { cn } from "@/lib/utils";
-import { buildPlatformSections, isNodeVisible } from "./tree";
+import { getVisibleUrlsByPlatform } from "./tree";
 
 const PLATFORM_OPTIONS: Record<
   Platform,
@@ -51,68 +51,17 @@ const PLATFORM_OPTIONS: Record<
   },
 };
 
-function isPathVisibleForPlatform(
-  visibleUrls: ReadonlySet<string>,
-  pathname: string,
-): boolean {
-  return visibleUrls.has(pathname);
-}
-
-function collectVisibleUrls(
-  node: PageTree.Node,
-  platform: Platform,
-  urls: Set<string>,
-): void {
-  if (!isNodeVisible(node, platform)) return;
-
-  if (node.type === "page") {
-    urls.add(node.url);
-    return;
-  }
-
-  if (node.type === "separator") return;
-
-  if (node.index && isNodeVisible(node.index, platform)) {
-    urls.add(node.index.url);
-  }
-
-  node.children.forEach((child) => {
-    collectVisibleUrls(child, platform, urls);
-  });
-}
-
-function getVisibleUrlsByPlatform(
-  tree: PageTree.Root | undefined,
-): Record<Platform, ReadonlySet<string>> {
-  const folders = (tree?.children ?? []).filter(
-    (node): node is PageTree.Folder => node.type === "folder",
-  );
-  const result: Record<Platform, Set<string>> = {
-    react: new Set(),
-    rn: new Set(),
-    ink: new Set(),
-  };
-
-  PLATFORMS.forEach((platform) => {
-    buildPlatformSections(folders, platform).forEach((section) => {
-      collectVisibleUrls(section, platform, result[platform]);
-    });
-  });
-
-  return result;
-}
-
 function getVisiblePlatformSwitchHref(
   visibleUrls: ReadonlySet<string>,
   pathname: string,
   nextPlatform: Platform,
 ): string {
   const equivalentHref = getPlatformSwitchHref(pathname, nextPlatform);
-  if (equivalentHref && isPathVisibleForPlatform(visibleUrls, equivalentHref)) {
+  if (equivalentHref && visibleUrls.has(equivalentHref)) {
     return equivalentHref;
   }
 
-  if (isPathVisibleForPlatform(visibleUrls, pathname)) {
+  if (visibleUrls.has(pathname)) {
     return pathname;
   }
